+++ src/eez/gui/widget.h
@@ -226,7 +226,7 @@ void drawWidgetCallback(const WidgetCursor &widgetCursor);
 
 OnTouchFunctionType getWidgetTouchFunction(const WidgetCursor &widgetCursor);
 
-uint16_t overrideStyleHook(const WidgetCursor &widgetCursor, uint16_t styleId);
+int overrideStyleHook(const WidgetCursor &widgetCursor, int styleId);
 uint16_t overrideStyleColorHook(const WidgetCursor &widgetCursor, const Style *style);
 uint16_t overrideActiveStyleColorHook(const WidgetCursor &widgetCursor, const Style *style);
